Add NewNote component and route
This commit involves creating a NewNote component that includes a form for new notes. A route for this component is also added to the Routes file. The MAX_ATTACHMENT_SIZE config property is also added for handling file attachments up to a certain size limit in the newly created note.